+++ weblate/fonts/tests/test_utils.py
@@ -4,6 +4,9 @@
 
 """Test for fonts."""
 
+from __future__ import annotations
+
+from django.core.cache import cache
 from django.test import SimpleTestCase
 
 from weblate.fonts.utils import check_render_size, get_font_weight
@@ -33,3 +36,25 @@ class RenderTest(SimpleTestCase):
                 lines=1,
             )
         )
+
+    def test_render_cache(self) -> None:
+        cache_key = "test:render:cache"
+
+        def invoke() -> bool:
+            return check_render_size(
+                font="sans",
+                weight=get_font_weight("normal"),
+                size=12,
+                spacing=0,
+                text="ahoj",
+                width=100,
+                lines=1,
+                cache_key=cache_key,
+            )
+
+        self.assertTrue(invoke())
+        cache.delete(cache_key)
+        self.assertFalse(cache.has_key(cache_key))
+
+        self.assertTrue(invoke())
+        self.assertTrue(cache.has_key(cache_key))